refactor(desktop): extract ChatPane types/constants/helpers (Phase 1/5)
Pure data + pure functions moved out of the 11.9k-line ChatPane.tsx
into focused sibling modules under ./ChatPane/. No runtime behavior
change; the orchestrator imports them back.

- types.ts        — all type/interface declarations
- constants.ts    — module-level constants, EMPTY_* sentinels,
                    storage keys, presets
- helpers.ts      — pure functions (slash/mention parsers, model
                    label formatters, time labels, attachment
                    type checks)
